refactor(server): renaming fold entities to Model, and BaseModels to BaseModel
