+++ server/src/models/FileProject.ts
@@ -2,12 +2,12 @@ import { IsBoolean, IsDate, IsString } from 'class-validator';
 import { Field, ObjectType } from 'type-graphql';
 import { Column, DeleteDateColumn, Entity, ManyToOne } from 'typeorm';
 
-import { Project } from 'entities/Project';
-import { BaseModels } from 'entities/base/BaseModels';
+import { Project } from 'models/Project';
+import { BaseModel } from 'models/base/BaseModel';
 
 @Entity()
 @ObjectType()
-export class FileProject extends BaseModels {
+export class FileProject extends BaseModel {
 	@Column()
 	@Field()
 	@IsString()
@@ -16,7 +16,7 @@ export class FileProject extends BaseModels {
 	@Column()
 	@Field()
 	@IsString()
-	code: string;
+	content: string;
 
 	@ManyToOne(() => Project, (project) => project.files)
 	project: Project;
