+++ server/src/models/ProgramingLanguage.ts
@@ -2,12 +2,12 @@ import { IsString } from 'class-validator';
 import { Field, ObjectType } from 'type-graphql';
 import { Column, Entity, Index, OneToMany } from 'typeorm';
 
-import { BaseModels } from 'entities/base/BaseModels';
-import { Project } from 'entities/Project';
+import { BaseModel } from 'models/base/BaseModel';
+import { Project } from 'models/Project';
 
 @Entity()
 @ObjectType()
-export class ProgramingLanguage extends BaseModels {
+export class ProgramingLanguage extends BaseModel {
 	@Column()
 	@Field()
 	@IsString()
